server/pyproject.toml
[project]
name = "game-server"
version = "1.0.0"
description = "游戏服务器 - 静态文件托管和 TTS 服务"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "dashscope>=1.17.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.0.0",
    "memorisdk>=2.1.0",
    "openai>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# 这是一个应用程序，包含所有 Python 文件
packages = ["."]

